Improve syntax for flattened()
Before:

    auto result1 =
      flattened(data, flatten_policy_discard<Key, Value, Interval>());
    auto result2 =
      flattened(data, FlattenPolicy<Key, Value, Interval>(func));

After:

    auto result1 = flattened(data, flatten_policy_discard());
    auto result2 = flattened(data, func);

Use structs with templated member function so that we don't
have to supply template type parameters when using
flatten_policy_discard() and flatten_policy_prefer_status_quo().

The <Key, Value, Interval> types will be deduced automatically in use.

flatten_policy_prefer_status_quo() is a special case because
it takes an "inner" fallback policy that obviously needs a type.
Strictly speaking, this should also be

    template <typename Key, typename Val, typename Interval>
    FlattenPolicy;

To avoid specifying <Key, Value, Interval> again, we save
the inner type as a generic

   template <typename FlattenPolicy>

All of this means that type checking is postponed until the
function is instantiated, and ugly compiler errors may ensue.
Concepts20 to the rescue later?