Add 'Rebased' trait to find type for joined_to() result
When using joined_to, we use a IntervalDict of A->B, B->C
and return an IntervalDict of A->C.

B->C does not need the same implementation as A->B.
However, the result (A->C) has the same implementation as
A->B but holding instances of "C" of course.

Rebased is a "type function" that when given the underlying
type for A->B, and the "C" type, will calculate the underlying
type for A->C based on the former.