Add Interval-tree implementation. `intervals()` results ordered by interval.
 - Interval Tree
  - Data is stored as contiguous, non-overlapping intervals per value.
  - Fast lookups with an interval tree
    (https://en.wikipedia.org/wiki/Interval_tree#Augmented_tree)
    where data is stored ordered by 1.interval 2.value
  - Manage inserts/deletes in a separate tree by ordering data
    by 1.value 2.intervals in order to split and merge intervals
    as necessary
  - Tree nodes shared between the two trees
 - `intervals(...)` results are sorted by interval then value rather than
   the other way around.

Interval tree code is currently from tinloaf/ygg library. This library supports
callbacks after internal tree operations using a traits class (much like
Boost::Graph). The version in interval_dict is a reimplementation of
ygg::interval_tree but instead of implicit open-closed intervals logic,
allows the full menagerie of intervals from Boost::icl.
Much of ygg traits-based logic abstracting away from the storage can be
simplified / removed since our data always uses boost::icl::intervals. Some
tricky logic is however required for retaining the original type of bounds for
dynamic (mixed open/closed) intervals after operations: if the original
interval was presented as open-closed or closed-closed, this should be returned
as such as far as possible.

Use boost::intrusive::set code for managing the same nodes in a red-black tree
ordered by 1.value and 2.intervals. Both trees are intrusive, allowing the
underlying nodes and data to be shared.

The overheads are a minimum of 6 pointers per interval, not including the
overheads of memory management, currently extravagantly using a hashtable.

The implementation code has been rewritten to use a traits class. See
 - interval_dict/adaptor.h
 - adaptor_icl_interval_map.h
 - adaptor_interval_tree.h
The implementations are now a dependent type of the container.
Function overload resolution is deferred to the second phase of template
instantiation. In practice this means that header order dependencies are no
longer fragile, and the implementation headers do not need to be included
before the containers. The abstract traits also provide a template for
future implementations.