Add CountOverlap: Overlapping subsequent intervals
Used in the implementation of augmented interval list

Count number of overlaps with subsequent intervals by sorting
the left and right edges together, and counting the number of
left edges before each right edge. Requires that left and
right edges are comparable, which is not true for intervals
that do not support boost::icl::first() and last()

In such cases, boost::icl::lower() and upper() are used instead
with resulting loss in accuracy.

Intervals with identical bounds are not considered overlapping.